Add script to guess stop_times.shape_dist_traveled
This patch converts the current untils to use SQLite as imported by cbrik's
gtfs_SQL_import project which is needed to efficiently process the
relationships involved in guessing the shape_dist_traveled on the stop_times
table. This patch also adds a script which does that guess work. It appears
to be reasonably accurate and sane.